[clang][modules] fmodules-driver C++20 module presence detection
This PR is part of a series aimed at implementing native support for
explicit module builds from the Clang driver.

This introduces a new scanner that detects C++20 module usage in
source files without using the preprocessor or lexer.
This will later help the Clang driver decide when to implicitly
perform an explicit module build.
Currently, it is only enabled with the `-fmodules-driver` flag and
meant for diagnostic purpose only.

The scanner is lightweight and is designed not to add noticeable
overhead to builds that don't use explicit module support. For more
details on performance, see:
https://gist.github.com/naveen-seth/a9f010dcd7ec7105c7b7205b544a6ddd

You full RFC/plan can be found here:
https://discourse.llvm.org/t/rfc-modules-support-simple-c-20-modules-use-from-the-clang-driver-without-a-build-system/86456/1